Allow to specify listen interfaces on Debian
The directive INTERFACESv4 in /etc/default/dhcp may be filled
with listening interfaces for the dhcp server. Hardcoding it to
ansible_default_ipv4.interface prevents the user to specify the
interface without extra tasks. It also affects idempotency on
this scenario. This modification adds an extra variable,
dhcp_debian_interfaces_v4, that defaults to ansible_default_ipv4.interface.
This way the behavior is maintained and the user can tweak it
if he wants for Debian OS family.