Allow to set permissions for custom includes
It is possible that some sensitive information to be put in the
custom include file, like a DNS updater secret key. In order
to guarantee apropriate permissions for the destination file
an extra variable is defined, dhcp_custom_includes_modes. The
filter zip_longest is used in conjunction with dhcp_custom_includes.
The default mode if not specified is 0644. A more elegant approach
would be to transform the dhcp_custom_includes variable in a
list of dicts instead of a list of strings. But that way would
break compatibility with previous versions.